+++ src/main/kotlin/com/iplease/lib/bento/api/error/data/ErrorData.kt
@@ -1,0 +1,3 @@
+package com.iplease.lib.bento.api.error.data
+
+interface ErrorData
