+++ src/main/kotlin/com/iplease/lib/bento/api/event/data/EventData.kt
@@ -1,0 +1,3 @@
+package com.iplease.lib.bento.api.event.data
+
+interface EventData
