+++ src/main/kotlin/com/iplease/lib/bento/api/event/type/EventType.kt
@@ -1,0 +1,9 @@
+package com.iplease.lib.bento.api.event.type
+
+import com.iplease.lib.bento.api.event.data.EventData
+import kotlin.reflect.KClass
+
+interface EventType {
+    fun getRoutingKey(): String
+    fun getEventDataType(): KClass<out EventData>
+}
