+++ src/main/kotlin/com/iplease/lib/bento/api/global/message/Message.kt
@@ -1,0 +1,6 @@
+package com.iplease.lib.bento.api.global.message
+
+interface Message {
+    fun getPayload(): ByteArray?
+    fun getRoutingKey(): String?
+}
