+++ src/main/kotlin/com/iplease/lib/bento/api/global/subscribe/MessageSubscriber.kt
@@ -1,0 +1,7 @@
+package com.iplease.lib.bento.api.global.subscribe
+
+import com.iplease.lib.bento.api.global.message.Message
+
+interface MessageSubscriber: Subscribable<Message> {
+    override fun subscribe(data: Message)
+}
