+++ src/main/kotlin/com/iplease/lib/bento/api/global/subscribe/Subscribable.kt
@@ -1,0 +1,5 @@
+package com.iplease.lib.bento.api.global.subscribe
+
+interface Subscribable<T: Any> {
+    fun subscribe(data: T)
+}
