--- src/main/kotlin/com/iplease/lib/bento/error/data/ErrorData.kt
@@ -1,3 +1,0 @@
-package com.iplease.lib.bento.error.data
-
-interface ErrorData
