--- src/main/kotlin/com/iplease/lib/bento/error/type/ErrorType.kt
@@ -1,9 +1,0 @@
-package com.iplease.lib.bento.error.type
-
-import com.iplease.lib.bento.error.data.ErrorData
-import kotlin.reflect.KClass
-
-interface ErrorType {
-    fun getRoutingKey(): String
-    fun getErrorDataType(): KClass<out ErrorData>
-}
