--- src/main/kotlin/com/iplease/lib/bento/error/type/GlobalErrorTypeV1.kt
@@ -1,15 +1,0 @@
-package com.iplease.lib.bento.error.type
-
-import com.iplease.lib.bento.error.data.ErrorData
-import kotlin.reflect.KClass
-
-enum class GlobalErrorTypeV1(
-    private val routingKey: String,
-    private val eventDataType: KClass<out ErrorData>
-): ErrorType {
-    WRONG_PAYLOAD("v1.error.wrong_payload", ErrorData::class),
-    UNKNOWN_ERROR("v1.error.unknown", ErrorData::class);
-
-    override fun getRoutingKey() = routingKey
-    override fun getErrorDataType() = eventDataType
-}
