--- src/main/kotlin/com/iplease/lib/bento/event/data/EventData.kt
@@ -1,3 +1,0 @@
-package com.iplease.lib.bento.event.data
-
-interface EventData
