--- src/main/kotlin/com/iplease/lib/bento/global/message/Message.kt
@@ -1,6 +1,0 @@
-package com.iplease.lib.bento.global.message
-
-interface Message {
-    fun getPayload(): ByteArray?
-    fun getRoutingKey(): String?
-}
