--- src/main/kotlin/com/iplease/lib/bento/global/subscribe/Subscribable.kt
@@ -1,5 +1,0 @@
-package com.iplease.lib.bento.global.subscribe
-
-interface Subscribable<T: Any> {
-    fun subscribe(data: T)
-}
